requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
opencv-Python
pymysql
azure-cognitiveservices-vision-face
numpy
azure-storage-blob==12.0.0